requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
dj-database-url==2.2.0
Django==5.1.1
gunicorn==23.0.0
idna==3.10
packaging==24.1
pillow==10.4.0
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
channels>=3.0
daphne>=3.0
